Add files via upload
FINAL PROJECT
Adams Andrew, the manager of WSDA Music, has been unable to account for a discrepancy in his company’s finances. The closest he’s gotten in his own attempts to analyse the data is figuring out that the discrepancy occurred between the years 2011 and 2012, but that’s about it. You have been called in to analyse WSDA’s data to get a list of names and then narrow down the list to pinpoint the prime suspects.